Fix typo in user registration complete and login, logout controller complete

+++ src/middleware/error.middleware.js
@@ -1,9 +1,13 @@
 import fs from 'fs'
 
 const errorMiddleware = (err, req, res, _next) => {
-    fs.unlinkSync(req.files.avatar[0].path)
-    fs.unlinkSync(req.files.coverImage[0].path)
 
+    if (req?.files?.coverImage[0].path) {
+        fs.unlinkSync(req.files.coverImage[0].path)
+    }
+    if (req?.files?.coverImage[0].path) {
+        fs.unlinkSync(req.files.coverImage[0].path)
+    }
     err.statusCode = err.statusCode || 500;
     err.message = err.message || "Error middleware issue"
